requirements.txt
flask-jwt-extended
flask-smorest
flask-sqlalchemy
passlib
marshmallow
python-dotenv
gunicorn
requests
pyyaml